% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Annotate_AOPs}
\alias{Annotate_AOPs}
\title{AOP_SSbD_ANNOTATION}
\format{
A data frame with 6 variables:
\describe{
\item{\code{AOP}}{AOP ids}
\item{\code{SSbD_category}}{SSbD impact category}
\item{\code{Endpoint}}{Endpoint}
\item{\code{Organ}}{Organ}
\item{\code{AOP_name}}{AOP_name}
\item{\code{Notes}}{Notes}
}
}
\source{
\url{https://doi.org/10.1038/s41597-023-02321-w}
}
\usage{
Annotate_AOPs
}
\description{
AOP_SSbD_IMPACT_ASSOCIATION
}
\keyword{datasets}
