% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aop_ke_table_hure}
\alias{aop_ke_table_hure}
\title{AOP_KE_ASSOCIATION}
\format{
A data frame with 5 variables:
\describe{
\item{\code{Aop}}{AOP ids}
\item{\code{Ke}}{KE ids}
\item{\code{Ke_type}}{KE type description, possible values are MolecularInitiatingEvent, AdverseOutcome, KeyEvent}
\item{\code{Ke_description}}{KE name}
\item{\code{a.name}}{AOP name}
}
}
\source{
\url{https://doi.org/10.1038/s41597-023-02321-w}
}
\usage{
aop_ke_table_hure
}
\description{
AOP_KE_MAPPING
}
\keyword{datasets}
