% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aop_names}
\alias{aop_names}
\title{AOP_NAMES}
\format{
A data frame with 2 variables:
\describe{
\item{\code{a.AOP_ID}}{AOP ids}
\item{\code{a.name}}{AOP names}
}
}
\source{
\url{https://doi.org/10.1038/s41597-023-02321-w}
}
\usage{
aop_names
}
\description{
AOP_IDs_NAMES_ASSOCIATION
}
\keyword{datasets}
