% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aops_new}
\alias{aops_new}
\title{AOP_KE_GENE_ASSOCIATION_DF}
\format{
A data frame with 6 variables:
\describe{
\item{\code{Aop}}{AOP ids}
\item{\code{Aop_KE}}{KE ids belonging to Aop id}
\item{\code{KE}}{KE id}
\item{\code{Annotation}}{GO id mapped to KE}
\item{\code{Level}}{Level}
\item{\code{Gene}}{Gene id}
}
}
\source{
\url{https://doi.org/10.1038/s41597-023-02321-w}
}
\usage{
aops_new
}
\description{
Dataframe of association between AOPs, KEs and ensembl genes
}
\keyword{datasets}
