% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Biological_system_annotations}
\alias{Biological_system_annotations}
\title{BIOLOGICAL_SYSTEM}
\format{
A data frame with 11 variables:
\describe{
\item{\code{a.AOP_ID}}{AOP ids}
\item{\code{a.name}}{AOP names}
\item{\code{key_event_name}}{Name of the key event in the biological system}
\item{\code{ke}}{Key event identifier}
\item{\code{level}}{Biological level of organization}
\item{\code{system}}{Primary biological system involved}
\item{\code{organ_tissue}}{Primary organ or tissue involved}
\item{\code{cell}}{Primary cell type involved}
\item{\code{cell_component}}{Primary cell component involved}
\item{\code{secondary_system}}{Secondary biological system involved, if any}
\item{\code{secondary_organ_tissue}}{Secondary organ or tissue involved, if any}
\item{\code{secondary_cell}}{Secondary cell type involved, if any}
\item{\code{secondary_cell_component}}{Secondary cell component involved, if any}
}
}
\source{
\url{https://doi.org/10.1038/s41597-023-02321-w}
}
\usage{
Biological_system_annotations
}
\description{
BIOLOGICAL_SYSTEM_ANNOTATION
}
\keyword{datasets}
