% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{build_aop_for_aop_fingeprints}
\alias{build_aop_for_aop_fingeprints}
\title{Build AOP Enrichment Results for AOP Fingerprints}
\usage{
build_aop_for_aop_fingeprints(
  aop_enrichment_results,
  ke_enrichment_results,
  min_aop_length = 6,
  percentage_enriched_ke = 0.33
)
}
\arguments{
\item{aop_enrichment_results}{Enrichment results for AOPs.}

\item{ke_enrichment_results}{Enrichment results for KEs.}

\item{min_aop_length}{Minimum length of AOPs. Default is 6.}

\item{percentage_enriched_ke}{Percentage of enriched KEs for an AOP to be considered significant. Default is 33\%.}

\item{time_var}{Variable representing time points.}
}
\value{
A list containing two data frames: detailed_results_only_enriched and detailed_results_all_ke_in_aop.
}
\description{
This function constructs AOP enrichment results for AOP fingerprints based on specified criteria.
}
