% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clust_anno}
\alias{clust_anno}
\title{CLUSTERED_KE}
\format{
A data frame with 3 variables:
\describe{
\item{\code{cluster}}{KE cluster ids}
\item{\code{KEs}}{KE ids}
\item{\code{Description}}{KE names}
}
}
\source{
\url{https://doi.org/10.1038/s41597-023-02321-w}
}
\usage{
clust_anno
}
\description{
KE-cluster associations
}
\keyword{datasets}
