% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEKE_network_functions.R
\name{compute_the_closest_AOs}
\alias{compute_the_closest_AOs}
\title{Compute closest adverse outcomes to a given vertex}
\usage{
compute_the_closest_AOs(
  interesting_vertex,
  KEKE_net,
  AO,
  max_path_length = 10,
  mode = "out",
  n_AOs = 20
)
}
\arguments{
\item{interesting_vertex}{A string representing the name of the vertex/vertices of interest in the KEKE_net.}

\item{KEKE_net}{An igraph object representing the knowledge exchange network.}

\item{AO}{A vector of adverse outcomes.}

\item{max_path_length}{An integer representing the maximum length of the path allowed to reach an adverse outcome (Including KE and AO).}

\item{n_AOs}{An integer representing the maximumn number adverse outcomes to be added}
}
\value{
the function returns a vector of the names of the closest AOs to the vertex of interest in the KEKE_net.
}
\description{
This function takes in a vertex (or set of vertices) of interest, an igraph object representing the KE-KE network,
a vector of adverse outcomes (AO), and optional parameters threshold and distance, and returns the closest adverse outcomes
to the vertex of interest in the KEKE_net. If distance is set to TRUE, it returns both the names of the closest adverse outcomes
and the corresponding distances between the vertex of interest and the adverse outcomes.
}
