% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEKE_network_functions.R
\name{compute_the_closest_MIEs}
\alias{compute_the_closest_MIEs}
\title{Compute closest molecular initiating events to a given vertex/set of vertices}
\usage{
compute_the_closest_MIEs(
  interesting_vertex,
  KEKE_net,
  MIE,
  max_path_length = 5,
  distance = TRUE,
  mode = "out",
  n_MIEs = 20
)
}
\arguments{
\item{interesting_vertex}{A string representing the name of the vertex/vertices of interest in the KEKE_net.}

\item{KEKE_net}{An igraph object representing the knowledge exchange network.}

\item{MIE}{A character vector representing the list of molecular initiating events (MIEs).}

\item{max_path_length}{An integer representing the maximum length of the path allowed (Including MIE and KE)}

\item{n_MIEs}{An integer representing the maximumn number adverse outcomes to be added#' @param distance A logical value (TRUE or FALSE) indicating whether to return both the names of the closest adverse outcomes and the corresponding distances (default is FALSE).}
}
\value{
the function returns a vector of the names of the closest MIEs to the vertex of interest in the KEKE_net.
}
\description{
This function takes in a vertex (or set of vertices) of interest, an igraph object representing the KE-KE network,
a vector of adverse outcomes (AO), and optional parameters threshold and distance, and returns the closest adverse outcomes
to the vertex of interest in the KEKE_net. If distance is set to TRUE, it returns both the names of the closest adverse outcomes
and the corresponding distances between the vertex of interest and the adverse outcomes.
}
