% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{convert_genes_to_symbol}
\alias{convert_genes_to_symbol}
\title{Convert a Vector of Gene IDs to Gene Symbols}
\usage{
convert_genes_to_symbol(
  genes,
  gene_id_type,
  organism,
  genes_human,
  genes_mouse,
  genes_rat
)
}
\arguments{
\item{genes}{A character vector of gene identifiers.}

\item{gene_id_type}{A string indicating the input gene ID type (e.g., \code{"ENSEMBL"}, \code{"ENTREZID"}, or \code{"SYMBOL"}).}

\item{organism}{A string indicating the organism (\code{"human"}, \code{"mouse"}, or \code{"rat"}).}

\item{genes_human}{mapping file for human genes}

\item{genes_mouse}{mapping file for mouse genes}

\item{genes_rat}{mapping file for rat genes}
}
\value{
A character vector of gene symbols.
}
\description{
Converts a vector of gene identifiers (e.g., ENSEMBL or ENTREZ IDs) to gene symbols for a specified organism.
Uses preloaded data frames (\code{genes_human}, \code{genes_mouse}, \code{genes_rat}) from \code{load_biomaRt_data()}.
}
