% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrich_KEs_AOPs}
\alias{enrich_KEs_AOPs}
\title{Perform Enrichment Analysis for Key Events}
\usage{
enrich_KEs_AOPs(
  GList,
  list_gene_sets,
  only_significant = FALSE,
  pval_th = 0.05,
  adj.method = "fdr",
  merge_by = "Ke",
  numerical_properties = "Score",
  background = unique(unlist(list_gene_sets)),
  verbose = TRUE,
  aggregation_function = "median"
)
}
\arguments{
\item{GList}{A named list of data frames. Each element of the list represents
a specific experimental condition. The dataframe must have a column with gene ids named 'Feature', and optional numerical
values associated to the genes (e.g. t-test statistics, fold-changes, bmd values, etc)}

\item{list_gene_sets}{One of the list of gene sets for enrichment analysis provided by the package}

\item{only_significant}{Logical indicating whether to include only significant results.}

\item{pval_th}{The p-value threshold for significance.}

\item{adj.method}{The method for p-value adjustment.}

\item{merge_by}{The variable for merging the results.}

\item{numerical_properties}{vector of strings, corresponding to colnames in the dataframes of the GList list that indicates numerical properties of the Features}

\item{background}{A vector of background genes to compare against.}

\item{verbose, }{boolean if TRUE a txtProgressBar is printed}

\item{aggregation_function}{Function to be used to aggregate the numerical values of the KEs/AOPs. Possible values are median, mean, min and quantile_05. When using quantile_05 the 5\% quantile of the distribution of the values is computed.}
}
\value{
A list of enriched key events (KEs) and associated results.
}
\description{
Conduct enrichment analysis for key events (KEs) using the provided data and parameters.
}
