% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_human}
\alias{genes_human}
\title{genes_human}
\format{
A dataframe with columns ensembl_gene_id, hgnc_symbol, entrezgene_id and description. From human = useMart("ensembl", dataset = "hsapiens_gene_ensembl", host = "https://dec2021.archive.ensembl.org/")
}
\usage{
genes_human
}
\description{
Mapping between ENSEMBL, SYMBOL and ENTREZ human gene IDs
}
\keyword{datasets}
