% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_mouse}
\alias{genes_mouse}
\title{genes_mouse}
\format{
A dataframe with columns ensembl_gene_id, hgnc_symbol, entrezgene_id and description. From mouse = useMart("ensembl", dataset = "mmusculus_gene_ensembl", host = "https://dec2021.archive.ensembl.org/")
}
\usage{
genes_mouse
}
\description{
Mapping between ENSEMBL, SYMBOL and ENTREZ mouse gene IDs
}
\keyword{datasets}
