% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_rat}
\alias{genes_rat}
\title{genes_rat}
\format{
A dataframe with columns ensembl_gene_id, hgnc_symbol, entrezgene_id and description.From  rat = useMart("ensembl", dataset = "rnorvegicus_gene_ensembl", host = "https://dec2021.archive.ensembl.org/")
}
\usage{
genes_rat
}
\description{
Mapping between ENSEMBL, SYMBOL and ENTREZ rat gene IDs
}
\keyword{datasets}
