% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{geneset_enrichment}
\alias{geneset_enrichment}
\title{Gene Set Enrichment Analysis}
\usage{
geneset_enrichment(
  genes_of_interest,
  background,
  genesets,
  adj.method = "fdr",
  verbose = FALSE,
  numerical_properties = c("BMD", "BMDU", "BMDL"),
  aggregation_function = "median"
)
}
\arguments{
\item{background}{A vector of background genes to compare against.}

\item{genesets}{A named list of gene sets, where each element is a vector of genes belonging to that set.}

\item{adj.method}{Method for adjusting p-values. Default is "fdr" (False Discovery Rate).}

\item{verbose}{Logical, indicating whether to print progress messages. Default is FALSE.}

\item{numerical_properties}{A vector of column names from the genes data frame representing numerical properties to be averaged. Default includes "BMD", "BMDU", and "BMDL".}

\item{aggregation_function}{Function to be used to aggregate the numerical values of the KEs/AOPs. Possible values are median, mean, min and quantile_05. When using quantile_05 the 5\% quantile of the distribution of the values is computed.}
}
\description{
This function performs a gene set enrichment analysis, comparing a list of genes to a reference set across multiple gene sets.
}
