% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{group_enriched_ke_by_tissue}
\alias{group_enriched_ke_by_tissue}
\title{Plot KEs average BMD by tissue}
\usage{
group_enriched_ke_by_tissue(
  ke_enrichment_results,
  filter_experiment = NULL,
  group_by_time = NULL,
  pheno_colnames = c("TermID", "Experiment", "organ_tissue", "BMD", "Ke_description"),
  relevant_tissues = c("liver", "kidney", "blood", "lung", "brain",
    "brain/nervous tissue", "brain/prefrontal cortex/liver")
)
}
\arguments{
\item{ke_enrichment_results}{A data frame containing the enrichment data.}

\item{filter_experiment}{a vector of strings of experiment IDs to be plotted. If NULL no filtering is performed}

\item{group_by_time}{a string for timeid. If NULL all time points are included in the plot. Use NULL also if timepoint is not a variable inclued in the experimental setup}

\item{pheno_colnames}{vector of string indicating the variables from KE enrichment results to be considered for plotting,}

\item{relevant_tissues}{list of tissues to add in the plot.}
}
\value{
A grouped bar plot
}
\description{
Plot KEs average BMD by tissue
}
