% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_entrez_aop}
\alias{human_entrez_aop}
\title{AOP_HUMAN_ENTREZ_GENES}
\format{
A list with AOP ids used as indices. Each position contains a vector of human gene symbol ids
}
\source{
\url{https://doi.org/10.1038/s41597-023-02321-w}
}
\usage{
human_entrez_aop
}
\description{
Mapping between AOP ids and human genes symbols
}
\keyword{datasets}
