% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ke_names}
\alias{ke_names}
\title{KE_NAMES}
\format{
A data frame with 2 variables:
\describe{
\item{\code{Ke}}{Ke ids}
\item{\code{Ke_description}}{KE names}
}
}
\source{
\url{https://doi.org/10.1038/s41597-023-02321-w}
}
\usage{
ke_names
}
\description{
KE_IDs_NAMES_ASSOCIATION
}
\keyword{datasets}
