% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KEKE_net}
\alias{KEKE_net}
\title{KE_KE_NETWORK}
\format{
It is an igraph object, whose nodes are the clustered KEs, and edges represent their connections. Nodes have the attribute 'ke_description' that includes a description of the KEs, while edges have the attribute 'type.r' that specifies the type of connections.
}
\source{
\url{https://doi.org/10.1038/s41597-023-02321-w}
}
\usage{
KEKE_net
}
\description{
KE-KE network.
}
\keyword{datasets}
