% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{loop_enrichment_v2}
\alias{loop_enrichment_v2}
\title{Enrichment Analysis for Multiple Experiment Genes}
\usage{
loop_enrichment_v2(
  GList,
  list_gene_sets,
  background,
  only_significant = TRUE,
  pval_th = 0.05,
  adj.method = "fdr",
  merge_by = "Ke",
  numerical_properties = c("Score"),
  verbose = TRUE,
  aggregation_function = "median"
)
}
\arguments{
\item{GList}{A named list of data frames. Each element of the list represents
a specific experimental condition. The dataframe must have a column with gene ids, and optional numerical
values associated to the genes (e.g. t-test statistics, fold-changes, bmd values, etc)}

\item{list_gene_sets}{A list of gene sets for enrichment analysis.}

\item{background}{Background gene set for enrichment analysis.}

\item{only_significant}{Logical, whether to include only significant results.}

\item{pval_th}{P-value threshold for significance.}

\item{adj.method}{Adjustment method for p-values.}

\item{merge_by}{Column used to merge results with the AOP-KE mapping table. Available options "Ke" or "Aop"}

\item{numerical_properties}{vector of strings, corresponding to colnames in the dataframes of the GList list that indicates numerical properties of the Features}

\item{verbose, }{boolean if TRUE a txtProgressBar is printed}

\item{aggregation_function}{Function to be used to aggregate the numerical values of the KEs/AOPs. Possible values are median, mean, min and quantile_05. When using quantile_05 the 5\% quantile of the distribution of the values is computed.}
}
\value{
A data frame containing the enriched results.
}
\description{
This is a wrapper of the function geneset_enrichment to performs enrichment analysis for multiple experiments.
}
