% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEKE_network_functions.R
\name{make_visNetwork}
\alias{make_visNetwork}
\title{Create a Visualization Network}
\usage{
make_visNetwork(
  detailed_results,
  experiment,
  enlarge_ke_selection = TRUE,
  ke_id,
  numerical_variables,
  pval_variable,
  gene_variable,
  convert_to_gene_symbols = F,
  gene_id_type,
  organism,
  max_path_length = 20,
  n_AOs = 10,
  n_MIEs = 10,
  mode = "out",
  genes_human,
  genes_mouse,
  genes_rat
)
}
\arguments{
\item{detailed_results}{A data frame containing detailed enrichment results. It should include columns corresponding to the experiment name, KE (Key Event) identifiers, KE descriptions, and other relevant numerical variables.}

\item{experiment}{The name of the experiment for which the network is being created. This should match the experiment name in the \code{detailed_results} data frame.}

\item{enlarge_ke_selection}{Logical, whether to expand the selection of Key Events (KEs) to include the closest Molecular Initiating Events (MIEs) and Adverse Outcomes (AOs). Default is TRUE.}

\item{ke_id}{The column name in \code{detailed_results} that corresponds to the Key Event (KE) identifier.}

\item{numerical_variables}{A character vector specifying the names of the columns in \code{detailed_results} that contain numerical variables to be included in the network nodes.}

\item{pval_variable}{The name of the column in \code{detailed_results} that contains p-values or adjusted p-values associated with the enrichment results.}

\item{gene_variable}{The name of the column in \code{detailed_results} that contains gene identifiers or gene sets associated with the enrichment results.}

\item{convert_to_gene_symbols}{Logical, whether to convert gene identifiers to gene symbols using organism-specific mapping. Default is FALSE.}

\item{gene_id_type}{A character string indicating the type of gene identifiers used (e.g., "ENSEMBL", "ENTREZID", or "SYMBOL").}

\item{organism}{A character string specifying the organism of interest ("human", "mouse", or "rat").}

\item{max_path_length}{Maximum length of the shortest path to be used to retrieve close MIEs or AOs if \code{enlarge_ke_selection = TRUE}. Default is 20.}

\item{n_AOs}{Number of AOs to add if \code{enlarge_ke_selection = TRUE}. Default is 10.}

\item{n_MIEs}{Number of MIEs to add if \code{enlarge_ke_selection = TRUE}. Default is 10.}

\item{mode}{Parameter passed to the \code{igraph::shortest.paths} function if \code{enlarge_ke_selection = TRUE}. Options are "all", "out", or "in". Default is "out".}

\item{genes_human}{mapping file for human genes}

\item{genes_mouse}{mapping file for mouse genes}

\item{genes_rat}{mapping file for rat genes}
}
\value{
A list containing two elements:
\describe{
\item{nodes}{A data frame of nodes representing the selected KEs, including information on the numerical variables, p-values, and associated genes.}
\item{edges}{A data frame of edges representing the connections between the KEs in the network.}
}
}
\description{
This function generates a visualization network (visNetwork) based on enrichment data and a network structure.
}
