% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_ens_aop}
\alias{mouse_ens_aop}
\title{AOP_MOUSE_ENSEMBL_GENES}
\format{
A list with AOP ids used as indices. Each position contains a vector of mouse ensembl gene ids
}
\source{
\url{https://doi.org/10.1038/s41597-023-02321-w}
}
\usage{
mouse_ens_aop
}
\description{
Mapping between AOP ids and mouse ensembl genes
}
\keyword{datasets}
