% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_ens_ke}
\alias{mouse_ens_ke}
\title{KE_MOUSE_ENSEMBL_GENES}
\format{
A list with KE ids used as indices. Each position contains a vector of mouse ensembl gene ids
}
\source{
\url{https://doi.org/10.1038/s41597-023-02321-w}
}
\usage{
mouse_ens_ke
}
\description{
Mapping between KE ids and mouse ensembl genes.
}
\keyword{datasets}
