% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_symbol_clusters}
\alias{mouse_symbol_clusters}
\title{CLUSTERED_KE_MOUSE_GENE_SYMBOLS}
\format{
A list with clustered KE ids used as indices. Each position contains a vector of mouse gene symbol ids
}
\source{
\url{https://doi.org/10.1038/s41597-023-02321-w}
}
\usage{
mouse_symbol_clusters
}
\description{
Mapping between KE ids and mouse gene symbols. KEs sharing the same set of genes are clustered in a combined id as KE_id1;KE_id2
}
\keyword{datasets}
