% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_symbol_ke}
\alias{mouse_symbol_ke}
\title{KE_MOUSE_GENE_SYMBOLS}
\format{
A list with KE ids used as indices. Each position contains a vector of mouse gene symbol ids
}
\source{
\url{https://doi.org/10.1038/s41597-023-02321-w}
}
\usage{
mouse_symbol_ke
}
\description{
Mapping between KE ids and mouse gene symbols.
}
\keyword{datasets}
