% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEKE_network_functions.R
\name{plot_visNetwork}
\alias{plot_visNetwork}
\title{Plot a Visualization Network}
\usage{
plot_visNetwork(
  nodes,
  edges,
  numerical_variables,
  group_by = "aop",
  color_enriched = "red",
  color_not_enriched = "gray"
)
}
\arguments{
\item{nodes}{A data frame containing nodes representing Key Events (KEs). Each node should include information such as KE identifiers, KE descriptions, p-values, associated genes, and any numerical variables relevant to the analysis.}

\item{edges}{A data frame containing edges representing the connections between the KEs in the network. Each edge should include information such as the source and target KE IDs.}

\item{numerical_variables}{A character vector specifying the names of the columns in \code{nodes} that contain numerical data to be displayed in the node tooltips.}

\item{group_by}{A character string indicating how to group the nodes in the network. Accepts either "aop" (default) to group by Adverse Outcome Pathways, or other specified categories such as "ssbd" to group by SSbD impact categories.Other grouping categories can be the colnames of the nodes parameter}
}
\value{
A visNetwork object that can be rendered to create an interactive network visualization. The network will color nodes based on their significance and group them according to the specified category.
}
\description{
This function generates a visualization network (visNetwork) based on the nodes and edges derived from enrichment data. It allows for the visual grouping of nodes based on either AOPs (Adverse Outcome Pathways) or SSbD (Safe and Sustainable by Design) impact categories, and highlights key events (KEs) based on their significance.
}
