% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rat_ens_ke}
\alias{rat_ens_ke}
\title{KE_RAT_ENSEMBL_GENES}
\format{
A list with KE ids used as indices. Each position contains a vector of rat ensembl gene ids
}
\source{
\url{https://doi.org/10.1038/s41597-023-02321-w}
}
\usage{
rat_ens_ke
}
\description{
Mapping between KE ids and rat ensembl genes.
}
\keyword{datasets}
