% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rat_entrez_aop}
\alias{rat_entrez_aop}
\title{AOP_RAT_ENTREZ_GENES}
\format{
A list with AOP ids used as indices. Each position contains a vector of rat gene symbol ids
}
\source{
\url{https://doi.org/10.1038/s41597-023-02321-w}
}
\usage{
rat_entrez_aop
}
\description{
Mapping between AOP ids and rat genes symbols
}
\keyword{datasets}
