% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{render_aop_fingerprint_bubble_plot}
\alias{render_aop_fingerprint_bubble_plot}
\title{Render AOP Fingerprint Bubble Plot}
\usage{
render_aop_fingerprint_bubble_plot(
  enrichement_data,
  group_AOPs = "SSbD_category",
  group_by,
  group_by2,
  x_axis_var,
  y_axis_var = "TermID",
  filter_column,
  filter_by,
  is_group_by_numeric,
  threshold_proportion,
  text_cex = 12
)
}
\arguments{
\item{enrichement_data}{A data frame containing the enrichment data.}

\item{group_by}{Variable used for grouping the data.}

\item{group_by2}{Second variable used for subgrouping the data.}

\item{filter_column}{Column for filtering the data.}

\item{filter_by}{Value for filtering the data.}

\item{is_group_by_numeric}{Logical, whether the grouping variable is numeric.}

\item{threshold_proportion}{Proportion threshold for filtering data.}

\item{text_cex}{Text size for labels.}

\item{time_var}{Column name of timepoint variable}
}
\value{
A bubble plot visualizing AOP fingerprint data.
}
\description{
This function generates a bubble plot for rendering AOP fingerprint data.
}
